Migration to the long format of id for transaction.

+++ client/src/main/java/com/orientechnologies/orient/client/remote/message/OBeginTransaction38Request.java
@@ -106,7 +106,7 @@ public class OBeginTransaction38Request implements OBinaryRequest<OBeginTransact
   @Override
   public void read(OChannelDataInput channel, int protocolVersion, ORecordSerializer serializer)
       throws IOException {
-    txId = channel.readInt();
+    txId = channel.readLong();
     hasContent = channel.readBoolean();
     usingLog = channel.readBoolean();
     operations = new ArrayList<>();
