+++ client/src/main/java/com/orientechnologies/orient/client/remote/message/OBeginTransactionRequest.java
@@ -95,7 +95,7 @@ public class OBeginTransactionRequest implements OBinaryRequest<OBeginTransactio
   @Override
   public void read(OChannelDataInput channel, int protocolVersion, ORecordSerializer serializer)
       throws IOException {
-    txId = channel.readInt();
+    txId = channel.readLong();
     hasContent = channel.readBoolean();
     usingLog = channel.readBoolean();
     operations = new ArrayList<>();
